+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,33 @@
+package com.techcourse.service;
+
+import com.interface21.transaction.TransactionManager;
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.domain.User;
+
+public class TxUserService implements UserService {
+
+    private final TransactionManager transactionManager;
+    private final UserService userService;
+
+    public TxUserService(UserService userService) {
+        this.userService = userService;
+        transactionManager = new TransactionManager(DataSourceConfig.getInstance());
+    }
+
+    @Override
+    public User findById(long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(User user) {
+        userService.save(user);
+    }
+
+    @Override
+    public void changePassword(long id, String newPassword, String createdBy) {
+        transactionManager.executeTransaction(connection -> {
+            userService.changePassword(id, newPassword, createdBy);
+        });
+    }
+}
