+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,47 +1,11 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.JdbcExecutionException;
-import com.interface21.transaction.TransactionManager;
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class UserService {
+public interface UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionManager transactionManager;
-
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        DataSource dataSource = DataSourceConfig.getInstance();
-        transactionManager = new TransactionManager(dataSource);
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final var user = findById(id);
-        user.changePassword(newPassword);
-        transactionManager.executeTransaction(connection -> {
-            userDao.update(connection, user);
-            userHistoryDao.log(connection, new UserHistory(user, createBy));
-        });
-    }
+    User findById(final long id);
+    void save(final User user);
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
